Compute the main formation and destruction routes of a given specie
If the trace_route keyword is set to 1 in the input file, Astrochem
now compute the rates of the main destruction and formation routes for
each species listed in "output". These rates, as well as the reaction
numbers, are written in .rout file files. A new program, plroute,
allow to plot these rates as a function of time. This allow to quickly
identify what are the important reactions in the formation or
destruction of a given specie.