Documentation changes
Among other changes in the documentation, figures are now created at
compile time by astrochem and plabun. A side change in plabun: it does
not anymore open a X11 window if the --output option is set. This
prevent for the abundances being plotted in a window when compiling
the documentation. The same change has been made in plroute for
consistency.