Plot abundances vs. Av rather than shell number
The plabun command has been modified so that it plots the abundances
as a function of the visual extinction rather than the shell
number. For this a --model option has been added to the command. This
option allows to specify the model (.mdl) file from which the Av
values are read.