New data structures and dynamic allocation
Code parameters have been regrouped into data structures, for a sake
of clarity. These data structures are dynamically allocated, so there
is no limit on the numbers of shells or reactions anymore. In addition
a new format of source model file has been implemented. This new
format allows to describe "time-dependant" sources, i.e. sources in
which the temperature, density, etc. changes as a function of
time. This new type of source has not implemented in solve() yet.