Fix libpyastrochem for Python 3
Strings in Python 3 are unicode by default, so the network file name
and the species names in the input abundance dictionary need to be
converted from unicode to C char in the Python interface. In
addition, species names in the returned abundances need to be
converted back from C char to Python strings. Closes #61.