Implement a function to set the initial abundance of a single species
This function is comparable to set_initial_abundances() which allow to
set the abundances of several species. But it is easier to wrap for
the Python, because it does not require to allocate an array of
strings. It is not useful outside the Python wrapper, so we don't
document it in the API.